app/requirements.txt
Django~=3.2
djangorestframework~=3.13
django-filter~=22.1
django-guardian~=2.4.0
psycopg2==2.9.5
gunicorn==20.1.0